+++ src/Cracking_the_coding_interview_five/eightQueens.java
@@ -1,4 +1,4 @@
-package Cracking_the_coding_interview_four;
+package Cracking_the_coding_interview_five;
 
 import java.util.ArrayList;
